@@ -1,12 +1,25 @@
 {
     "name": "papelo",
-    "author": {
-        "name": "im-parsa",
-        "email": "contact@parsa-firoozi.ir",
-        "url": "https://parsa-firoozi.ir"
-    },
-    "version": "1.0.0",
+    "version": "0.1.0",
     "private": true,
+    "author": "IntelligentQuantum - Im-Parsa",
+    "dependencies": {
+        "@testing-library/jest-dom": "^5.16.4",
+        "@testing-library/react": "^13.3.0",
+        "@testing-library/user-event": "^13.5.0",
+        "@types/jest": "^27.5.2",
+        "@types/node": "^16.11.39",
+        "@types/react": "^18.0.12",
+        "@types/react-dom": "^18.0.5",
+        "classnames": "^2.3.1",
+        "react": "^18.1.0",
+        "react-dom": "^18.1.0",
+        "react-router-dom": "^6.3.0",
+        "react-scripts": "5.0.1",
+        "sass": "^1.52.3",
+        "typescript": "^4.7.3",
+        "web-vitals": "^2.1.4"
+    },
     "scripts": {
         "start": "react-scripts start",
         "build": "react-scripts build",
@@ -30,32 +43,5 @@
             "last 1 firefox version",
             "last 1 safari version"
         ]
-    },
-    "devDependencies": {
-        "@types/axios": "^0.14.0",
-        "@types/redux-logger": "^3.0.9",
-        "@types/styled-components": "^5.1.20"
-    },
-    "dependencies": {
-        "@testing-library/jest-dom": "^5.16.1",
-        "@testing-library/react": "^12.1.2",
-        "@testing-library/user-event": "^13.5.0",
-        "@tippyjs/react": "^4.2.6",
-        "@types/jest": "^27.4.0",
-        "@types/react": "^17.0.38",
-        "@types/react-dom": "^17.0.11",
-        "axios": "^0.25.0",
-        "moment": "^2.29.1",
-        "react": "^17.0.2",
-        "react-dom": "^17.0.2",
-        "react-redux": "^7.2.6",
-        "react-router-dom": "^6.2.1",
-        "react-scripts": "5.0.0",
-        "redux": "^4.1.2",
-        "redux-logger": "^3.0.6",
-        "sass": "^1.48.0",
-        "styled-components": "^5.3.5",
-        "typescript": "^4.5.4",
-        "web-vitals": "^2.1.3"
     }
 }
