@@ -1,13 +1,13 @@
 import React from 'react';
 import { Route, Routes } from 'react-router-dom';
 
-import Home from './pages/home/home.components';
+import Home from "./routes/home/home.component";
 
-const App = () =>
+function App()
 {
     return (
         <Routes>
-            <Route path='/' element={<Home/>}/>
+            <Route path='/' element={ <Home /> }/>
         </Routes>
     );
 }
