@@ -4,7 +4,7 @@ const reportWebVitals = (onPerfEntry?: ReportHandler) =>
 {
     if (onPerfEntry && onPerfEntry instanceof Function)
     {
-        import('web-vitals').then(({getCLS, getFID, getFCP, getLCP, getTTFB}) =>
+        import('web-vitals').then(({ getCLS, getFID, getFCP, getLCP, getTTFB }) =>
         {
             getCLS(onPerfEntry);
             getFID(onPerfEntry);
